Revert "add mysql"
