+++ cds/Provider.go
@@ -46,16 +46,15 @@ func Provider() terraform.ResourceProvider {
 			"cds_data_source_private_subnet": dataSourceCdsPrivateSubnet(),
 			"cds_data_source_haproxy":        dataSourceHaproxy(),
 			"cds_data_source_certificate":    dataSourceHaproxyCertificate(),
-			"cds_data_source_mysql":          dataSourceCdsMySQL(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
-			"cds_vdc":            resourceCdsVdc(),
-			"cds_instance":       resourceCdsCcsInstance(),
-			"cds_security_group": resourceCdsSecurityGroup(),
-			"cds_private_subnet": resourceCdsPrivateSubnet(),
-			"cds_haproxy":        resourceCdsHaproxy(),
-			"cds_certificate":    resourceCdsCert(),
-			"cds_mysql":          resourceCdsMySQL(),
+			"cds_vdc":              resourceCdsVdc(),
+			"cds_instance":         resourceCdsCcsInstance(),
+			"cds_security_group":   resourceCdsSecurityGroup(),
+			"cds_private_subnet":   resourceCdsPrivateSubnet(),
+			"cds_haproxy":          resourceCdsHaproxy(),
+			"cds_haproxy_strategy": resourceCdsHaproxyStrategy(),
+			"cds_certificate":      resourceCdsCert(),
 		},
 
 		ConfigureFunc: providerConfigure,
