+++ cds/resource_cds_haproxy.go
@@ -2,10 +2,9 @@ package cds
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"log"
+	"strconv"
 	"strings"
 	"time"
 
@@ -21,6 +20,11 @@ func resourceCdsHaproxy() *schema.Resource {
 		Update: updateResourceCdsHaproxy,
 		Delete: deleteRresourceCdsHaproxy,
 		Schema: map[string]*schema.Schema{
+			"instance_uuid": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Instance UUID.",
+			},
 			"region_id": {
 				Type:        schema.TypeString,
 				Required:    true,
@@ -41,15 +45,10 @@ func resourceCdsHaproxy() *schema.Resource {
 				Required:    true,
 				Description: "instance name.",
 			},
-			"cpu": {
+			"paas_goods_id": {
 				Type:        schema.TypeInt,
 				Required:    true,
-				Description: "instance cpu num",
-			},
-			"ram": {
-				Type:        schema.TypeInt,
-				Required:    true,
-				Description: "instance ram size",
+				Description: "pass goods id.",
 			},
 			"ips": {
 				Type:       schema.TypeList,
@@ -72,194 +71,6 @@ func resourceCdsHaproxy() *schema.Resource {
 					},
 				},
 			},
-			"http_listeners": {
-				Type:        schema.TypeList,
-				ConfigMode:  schema.SchemaConfigModeAttr,
-				Optional:    true,
-				Description: "http listeners",
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"acl_white_list": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"backend_server": {
-							Type:       schema.TypeList,
-							Optional:   true,
-							ConfigMode: schema.SchemaConfigModeAttr,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"ip": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-									"max_conn": {
-										Type:     schema.TypeInt,
-										Required: true,
-									},
-									"port": {
-										Type:     schema.TypeInt,
-										Required: true,
-									},
-									"weight": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-								},
-							},
-						},
-						"certificate_ids": {
-							Type:       schema.TypeList,
-							Optional:   true,
-							Computed:   true,
-							ConfigMode: schema.SchemaConfigModeAttr,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"certificate_id": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-									"certificate_name": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-								},
-							},
-						},
-						"client_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"client_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"connect_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"connect_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_mode": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_name": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_port": {
-							Type:     schema.TypeInt,
-							Optional: true,
-						},
-						"max_conn": {
-							Type:     schema.TypeInt,
-							Optional: true,
-						},
-						"scheduler": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"server_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"server_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"sticky_session": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-					},
-				},
-			},
-			"tcp_listeners": {
-				Type:        schema.TypeList,
-				Optional:    true,
-				Description: "base pipe id.",
-				ConfigMode:  schema.SchemaConfigModeAttr,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"acl_white_list": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"backend_server": {
-							Type:       schema.TypeList,
-							Optional:   true,
-							ConfigMode: schema.SchemaConfigModeAttr,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"ip": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-									"max_conn": {
-										Type:     schema.TypeInt,
-										Required: true,
-									},
-									"port": {
-										Type:     schema.TypeInt,
-										Required: true,
-									},
-									"weight": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-								},
-							},
-						},
-						"client_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"client_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"connect_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"connect_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_mode": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_name": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"listener_port": {
-							Type:     schema.TypeInt,
-							Optional: true,
-						},
-						"max_conn": {
-							Type:     schema.TypeInt,
-							Optional: true,
-						},
-						"scheduler": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"server_timeout": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-						"server_timeout_unit": {
-							Type:     schema.TypeString,
-							Optional: true,
-						},
-					},
-				},
-			},
 		},
 	}
 }
@@ -274,31 +85,12 @@ func createResourceCdsHaproxy(data *schema.ResourceData, meta interface{}) error
 
 	request := haproxy.NewCreateLoadBalancerRequest()
 
-	var cpu, ram int
-	var region string
-
 	if inter, ok := data.GetOk("region_id"); ok {
 		regionId, exist := inter.(string)
 		if exist {
 			request.RegionId = common.StringPtr(regionId)
 		}
-		region = regionId
 	}
-
-	if inter, ok := data.GetOk("cpu"); ok {
-		cpu, _ = inter.(int)
-	}
-
-	if inter, ok := data.GetOk("ram"); ok {
-		ram, _ = inter.(int)
-	}
-
-	paasGoodsId, err := matchPassGoodsId(ctx, haproxyService, cpu, ram, region)
-	if err != nil {
-		return err
-	}
-	request.PaasGoodsId = common.IntPtr(paasGoodsId)
-
 	if inter, ok := data.GetOk("vdc_id"); ok {
 		vdcId, exist := inter.(string)
 		if exist {
@@ -317,7 +109,12 @@ func createResourceCdsHaproxy(data *schema.ResourceData, meta interface{}) error
 			request.InstanceName = common.StringPtr(instanceName)
 		}
 	}
-
+	if inter, ok := data.GetOk("paas_goods_id"); ok {
+		paasGoodsId, exist := inter.(int)
+		if exist {
+			request.PaasGoodsId = common.IntPtr(paasGoodsId)
+		}
+	}
 	if inter, ok := data.GetOk("ips"); ok {
 		ips, exist := inter.([]interface{})
 		if exist {
@@ -343,109 +140,32 @@ func createResourceCdsHaproxy(data *schema.ResourceData, meta interface{}) error
 	}
 	request.Amount = common.IntPtr(1)
 
-	resp, err := haproxyService.CreateHaproxy(ctx, request)
+	_, err := haproxyService.CreateHaproxy(ctx, request)
 	if err != nil {
 		return err
 	}
 
-	if len(resp.Data.InstancesUuid) == 0 {
-		return errors.New("create haproxy failed, please check")
-	}
+	data.Set("instance_uuid", "")
 
-	instanceUuid := resp.Data.InstancesUuid[0]
+	data.SetId("Need to Set Instance UUID" + strconv.FormatInt(time.Now().Unix(), 10))
 
-	data.SetId(instanceUuid)
+	time.Sleep(time.Second * 30)
 
-	if err := waitHaproxyRunning(ctx, haproxyService, instanceUuid); err != nil {
-		return err
-	}
+	// TODO wait until task id has effect
+	// taskService := TaskService{client: meta.(*CdsClient).apiConn}
+	// detail, err := taskService.DescribeTask(ctx, response.TaskId)
+	// if err != nil {
+	// 	return err
+	// }
 
-	// 创建策略
-	strategyRequest := haproxy.NewModifyLoadBalancerStrategysRequest()
-	strategyRequest.InstanceUuid = &instanceUuid
+	// if *detail.Code != "Success" {
+	// 	return fmt.Errorf("[ERROR] create haproxy task info get failed, error: %s", *detail.Message)
+	// }
 
-	httpListeners := make([]*HaproxyStrategyHttpListenerProviderInput, 0)
-	bytesData, _ := json.Marshal(data.Get("http_listeners"))
-	if err := json.Unmarshal(bytesData, &httpListeners); err != nil {
-		return err
-	}
-	for _, httpListener := range httpListeners {
-		backendServer := []*haproxy.DescribeLoadBalancerStrategysBackendServer{}
-		for _, backendServerEntry := range httpListener.BackendServer {
-			backendServer = append(backendServer, &haproxy.DescribeLoadBalancerStrategysBackendServer{
-				IP:      &backendServerEntry.IP,
-				MaxConn: &backendServerEntry.MaxConn,
-				Port:    &backendServerEntry.Port,
-				Weight:  &backendServerEntry.Weight,
-			})
-		}
-		httpListenerEntry := &haproxy.DescribeLoadBalancerStrategysHttpListeners{
-			ServerTimeoutUnit:  &httpListener.ServerTimeoutUnit,
-			ServerTimeout:      &httpListener.ServerTimeout,
-			StickySession:      &httpListener.StickySession,
-			AclWhiteList:       common.StringPtrs(strings.Split(strings.TrimSpace(httpListener.AclWhiteList), ",")),
-			ListenerMode:       &httpListener.ListenerMode,
-			MaxConn:            &httpListener.MaxConn,
-			ConnectTimeout:     &httpListener.ConnectTimeout,
-			ConnectTimeoutUnit: &httpListener.ConnectTimeoutUnit,
-			Scheduler:          &httpListener.Scheduler,
-			ClientTimeout:      &httpListener.ClientTimeout,
-			ClientTimeoutUnit:  &httpListener.ClientTimeoutUnit,
-			ListenerName:       &httpListener.ListenerName,
-			ListenerPort:       &httpListener.ListenerPort,
-			BackendServer:      backendServer,
-		}
-
-		strategyRequest.HttpListeners = append(strategyRequest.HttpListeners, httpListenerEntry)
-	}
-
-	tcpListeners := make([]*HaproxyStrategyTcpListenerProviderInput, 0)
-	bytesData, _ = json.Marshal(data.Get("tcp_listeners"))
-	if err := json.Unmarshal(bytesData, &tcpListeners); err != nil {
-		return err
-	}
-
-	for _, tcpListener := range tcpListeners {
-		backendServer := []*haproxy.DescribeLoadBalancerStrategysBackendServer{}
-		for _, backendServerEntry := range tcpListener.BackendServer {
-			backendServer = append(backendServer, &haproxy.DescribeLoadBalancerStrategysBackendServer{
-				IP:      &backendServerEntry.IP,
-				MaxConn: &backendServerEntry.MaxConn,
-				Port:    &backendServerEntry.Port,
-				Weight:  &backendServerEntry.Weight,
-			})
-		}
-		tcpListenerEntry := &haproxy.DescribeLoadBalancerStrategysTcpListeners{
-			ServerTimeoutUnit:  &tcpListener.ServerTimeoutUnit,
-			ServerTimeout:      &tcpListener.ServerTimeout,
-			AclWhiteList:       common.StringPtrs(strings.Split(strings.TrimSpace(tcpListener.AclWhiteList), ",")),
-			ListenerMode:       &tcpListener.ListenerMode,
-			MaxConn:            &tcpListener.MaxConn,
-			ConnectTimeout:     &tcpListener.ConnectTimeout,
-			ConnectTimeoutUnit: &tcpListener.ConnectTimeoutUnit,
-			Scheduler:          &tcpListener.Scheduler,
-			ClientTimeout:      &tcpListener.ClientTimeout,
-			ClientTimeoutUnit:  &tcpListener.ClientTimeoutUnit,
-			ListenerName:       &tcpListener.ListenerName,
-			ListenerPort:       &tcpListener.ListenerPort,
-			BackendServer:      backendServer,
-		}
-		strategyRequest.TcpListeners = append(strategyRequest.TcpListeners, tcpListenerEntry)
-	}
-
-	response, err := haproxyService.ModifyHaproxyStrategy(ctx, strategyRequest)
-	if err != nil {
-		return err
-	}
-	if *response.Code != "Success" {
-		return fmt.Errorf("Haproxy modify haproxy strategy with error:" + err.Error())
-	}
-
-	if err := waitHaproxyRunning(ctx, haproxyService, instanceUuid); err != nil {
-		return err
-	}
-
-	return readResourceCdsHaproxy(data, meta)
+	// data.SetId(strings.Join(common.StringValues(detail.Data.ResourceIds), ","))
+	// time.Sleep(time.Second * 10)
+	// return readResourceCdsHaproxy(data, meta)
+	return nil
 }
 
 func readResourceCdsHaproxy(data *schema.ResourceData, meta interface{}) error {
@@ -463,52 +183,32 @@ func updateResourceCdsHaproxy(data *schema.ResourceData, meta interface{}) error
 
 	haproxyService := HaproxyService{client: meta.(*CdsClient).apiConn}
 
-	var cpu, ram int
-	var hasChange bool
-
-	if data.HasChange("cpu") {
-		hasChange = true
-		inter, _ := data.GetOk("cpu")
-		cpu = inter.(int)
+	if data.HasChange("instance_uuid") {
+		data.SetId(strconv.FormatInt(time.Now().Unix(), 10))
 	}
 
-	if data.HasChange("ram") {
-		hasChange = true
-		inter, _ := data.GetOk("ram")
-		ram = inter.(int)
-	}
+	if !strings.Contains(data.Id(), "Need to Set Instance UUID") {
+		if data.HasChanges("paas_goods_id") {
 
-	inter, _ := data.GetOk("region_id")
-	regionId := inter.(string)
+			request := haproxy.NewModifyLoadBalancerInstanceSpecRequest()
+			if paasGoodsId, ok := data.GetOk("paas_goods_id"); ok {
+				request.PaasGoodsId = common.IntPtr(paasGoodsId.(int))
+			}
 
-	if hasChange {
-		request := haproxy.NewModifyLoadBalancerInstanceSpecRequest()
-		paasGoodsId, err := matchPassGoodsId(ctx, haproxyService, cpu, ram, regionId)
-		if err != nil {
-			return err
+			if instanceUUID, ok := data.GetOk("instance_uuid"); ok {
+				request.InstanceUuid = common.StringPtr(instanceUUID.(string))
+			}
+
+			_, err := haproxyService.ModifyHaproxy(ctx, request)
+			if err != nil {
+				return err
+			}
 		}
-
-		request.InstanceUuid = common.StringPtr(data.Id())
-		request.PaasGoodsId = &paasGoodsId
-
-		_, err = haproxyService.ModifyHaproxy(ctx, request)
-		if err != nil {
-			return err
-		}
-
-		if err := waitHaproxyRunning(ctx, haproxyService, data.Id()); err != nil {
-			return err
-		}
+	} else {
+		return errors.New("You must update instance_uuid in cds_haproxy block by your self, it will fix in the future")
 	}
 
-	modifyResponse, err := haproxyService.ModifyHaproxyStrategy(ctx, createModitifyStrategyRequest(data))
-	if err != nil {
-		return err
-	}
-
-	if *modifyResponse.Code != "Success" {
-		return errors.New(*modifyResponse.Message)
-	}
+	time.Sleep(time.Second * 5)
 
 	return readResourceCdsHaproxy(data, meta)
 }
@@ -519,295 +219,206 @@ func deleteRresourceCdsHaproxy(data *schema.ResourceData, meta interface{}) erro
 	logId := getLogId(contextNil)
 	ctx := context.WithValue(context.TODO(), "logId", logId)
 
-	haproxyService := HaproxyService{client: meta.(*CdsClient).apiConn}
+	if !strings.Contains(data.Id(), "Need to Set Instance UUID") {
+		haproxyService := HaproxyService{client: meta.(*CdsClient).apiConn}
 
-	request := haproxy.NewDeleteLoadBalancerRequest()
-	request.InstanceUuid = common.StringPtr(data.Id())
+		request := haproxy.NewDeleteLoadBalancerRequest()
+		if instanceUUID, ok := data.GetOk("instance_uuid"); ok {
+			request.InstanceUuid = common.StringPtr(instanceUUID.(string))
+		}
 
-	_, err := haproxyService.DeleteHaproxy(ctx, request)
-	if err != nil {
-		return err
-	}
+		_, err := haproxyService.DeleteHaproxy(ctx, request)
+		if err != nil {
+			return err
+		}
 
-	if err := waitHaproxyDelete(ctx, haproxyService, data.Id()); err != nil {
-		return err
+	} else {
+		return errors.New("You must update instance_uuid in cds_haproxy block by your self, it will fix in the future")
 	}
 	return nil
-}
-
-func waitHaproxyRunning(ctx context.Context, service HaproxyService, instanceUuid string) error {
-	descRequest := haproxy.NewDescribeLoadBalancersRequest()
-	descRequest.InstanceUuid = &instanceUuid
-
-	// 等待直到创建成功返回
-	for {
-		time.Sleep(time.Second * 15)
-		descResp, err := service.DescribeHaproxy(ctx, descRequest)
-		if err != nil {
-			return err
-		}
-
-		if *descResp.Code != "Success" {
-			return errors.New(*descResp.Message)
-		}
-
-		for _, data := range descResp.Data {
-			if *data.Status == "RUNNING" {
-				return nil
-			}
-		}
-	}
-}
-
-func waitHaproxyDelete(ctx context.Context, service HaproxyService, instanceUuid string) error {
-	descRequest := haproxy.NewDescribeLoadBalancersRequest()
-	descRequest.InstanceUuid = &instanceUuid
-
-	for {
-		time.Sleep(time.Second * 15)
-		descResp, err := service.DescribeHaproxy(ctx, descRequest)
-		if err != nil {
-			return err
-		}
-
-		if *descResp.Code == "RESOURCE_NOT_FOUND" {
-			return nil
-		}
-	}
-}
-
-func matchPassGoodsId(ctx context.Context, service HaproxyService, cpu, ram int, regionId string) (int, error) {
-	goodsRequest := haproxy.NewDescribeLoadBalancersSpecRequest()
-
-	goodsRequest.RegionId = common.StringPtr(regionId)
-
-	goodsResponse, err := service.DescribeLoadBalancersSpec(ctx, goodsRequest)
-	if err != nil {
-		return -1, err
-	}
-
-	for _, product := range goodsResponse.Data.Products {
-		for _, arch := range product.Architectures {
-			for _, role := range arch.ComputeRoles {
-				for _, cpuRam := range role.Standards.CpuRam {
-					if *cpuRam.CPU == cpu && *cpuRam.RAM == ram {
-						return *cpuRam.PaasGoodsId, nil
-					}
-				}
-			}
-		}
-	}
-
-	return -1, fmt.Errorf("cpu %d, ram %d not found paas_goods_id", cpu, ram)
 }
 
 func createModitifyStrategyRequest(data *schema.ResourceData) *haproxy.ModifyLoadBalancerStrategysRequest {
 	request := haproxy.NewModifyLoadBalancerStrategysRequest()
 
-	request.InstanceUuid = common.StringPtr(data.Id())
-
-	if inter, ok := data.GetOk("http_listeners"); ok {
-		datas := inter.([]interface{})
-		listeners := make([]*haproxy.DescribeLoadBalancerStrategysHttpListeners, 0)
-		for _, data := range datas {
-			if data == nil {
-				continue
-			}
-			dataMap := data.(map[string]interface{})
-
-			backendServers := make([]*haproxy.DescribeLoadBalancerStrategysBackendServer, 0)
-			if dataMap["backend_server"] != nil {
-				backendServersDatas := dataMap["backend_server"].([]interface{})
-				for _, backendServersData := range backendServersDatas {
-					backendServerMap := backendServersData.(map[string]interface{})
-					backendServers = append(backendServers, &haproxy.DescribeLoadBalancerStrategysBackendServer{
-						IP:      common.StringPtr(backendServerMap["ip"].(string)),
-						MaxConn: common.IntPtr(backendServerMap["max_conn"].(int)),
-						Port:    common.IntPtr(backendServerMap["port"].(int)),
-						Weight:  common.StringPtr(backendServerMap["weight"].(string)),
-					})
-				}
-			}
-
-			certificateIds := make([]*haproxy.DescribeLoadBalancerStrategysCertificateIds, 0)
-			if dataMap["certificate_ids"] != nil {
-				certificateIdsDatas := dataMap["certificate_ids"].([]interface{})
-				for _, certificateIdsData := range certificateIdsDatas {
-					certificateIdsMap := certificateIdsData.(map[string]interface{})
-					certificateIds = append(certificateIds, &haproxy.DescribeLoadBalancerStrategysCertificateIds{
-						CertificateId:   common.StringPtr(certificateIdsMap["certificate_id"].(string)),
-						CertificateName: common.StringPtr(certificateIdsMap["certificate_name"].(string)),
-					})
-				}
-			} else {
-				certificateIds = []*haproxy.DescribeLoadBalancerStrategysCertificateIds{}
-			}
-
-			listener := &haproxy.DescribeLoadBalancerStrategysHttpListeners{
-				BackendServer:  backendServers,
-				CertificateIds: certificateIds,
-			}
-
-			if dataMap["acl_white_list"] != nil {
-				listener.AclWhiteList = common.StringPtrs(strings.Split(dataMap["acl_white_list"].(string), ","))
-			} else {
-				listener.AclWhiteList = []*string{}
-			}
-			if dataMap["client_timeout"] != nil {
-				listener.ClientTimeout = common.StringPtr(dataMap["client_timeout"].(string))
-			}
-			if dataMap["client_timeout_unit"] != nil {
-				listener.ClientTimeoutUnit = common.StringPtr(dataMap["client_timeout_unit"].(string))
-			}
-			if dataMap["connect_timeout"] != nil {
-				listener.ConnectTimeout = common.StringPtr(dataMap["connect_timeout"].(string))
-			}
-			if dataMap["connect_timeout_unit"] != nil {
-				listener.ConnectTimeoutUnit = common.StringPtr(dataMap["connect_timeout_unit"].(string))
-			}
-			if dataMap["listener_mode"] != nil {
-				listener.ListenerMode = common.StringPtr(dataMap["listener_mode"].(string))
-			}
-			if dataMap["listener_name"] != nil {
-				listener.ListenerName = common.StringPtr(dataMap["listener_name"].(string))
-			}
-			if dataMap["listener_port"] != nil {
-				listener.ListenerPort = common.IntPtr(dataMap["listener_port"].(int))
-			}
-			if dataMap["max_conn"] != nil {
-				listener.MaxConn = common.IntPtr(dataMap["max_conn"].(int))
-			}
-			if dataMap["scheduler"] != nil {
-				listener.Scheduler = common.StringPtr(dataMap["scheduler"].(string))
-			}
-			if dataMap["server_timeout"] != nil {
-				listener.ServerTimeout = common.StringPtr(dataMap["server_timeout"].(string))
-			}
-			if dataMap["server_timeout_unit"] != nil {
-				listener.ServerTimeoutUnit = common.StringPtr(dataMap["server_timeout_unit"].(string))
-			}
-			if dataMap["sticky_session"] != nil {
-				listener.StickySession = common.StringPtr(dataMap["sticky_session"].(string))
-			}
-			listeners = append(listeners, listener)
+	if inter, ok := data.GetOk("instance_uuid"); ok {
+		instanceUuid, exist := inter.(string)
+		if exist {
+			request.InstanceUuid = common.StringPtr(instanceUuid)
 		}
-		request.HttpListeners = listeners
 	}
 
-	if inter, ok := data.GetOk("tcp_listeners"); ok {
-		datas := inter.([]interface{})
-		listeners := make([]*haproxy.DescribeLoadBalancerStrategysTcpListeners, 0)
-		for _, data := range datas {
-			if data == nil {
-				continue
-			}
-			dataMap := data.(map[string]interface{})
+	inter, ok := data.GetOk("strategies")
+	if !ok {
+		return request
+	}
+	strategies, ok := inter.([]interface{})
+	if !ok {
+		return request
+	}
 
-			backendServers := make([]*haproxy.DescribeLoadBalancerStrategysBackendServer, 0)
-			if dataMap["backend_server"] != nil {
-				backendServersDatas := dataMap["backend_server"].([]interface{})
-				for _, backendServersData := range backendServersDatas {
-					backendServerMap := backendServersData.(map[string]interface{})
-					backendServers = append(backendServers, &haproxy.DescribeLoadBalancerStrategysBackendServer{
-						IP:      common.StringPtr(backendServerMap["ip"].(string)),
-						MaxConn: common.IntPtr(backendServerMap["max_conn"].(int)),
-						Port:    common.IntPtr(backendServerMap["port"].(int)),
-						Weight:  common.StringPtr(backendServerMap["weight"].(string)),
-					})
-				}
-			}
-
-			listener := &haproxy.DescribeLoadBalancerStrategysTcpListeners{
-				BackendServer: backendServers,
-			}
-
-			if dataMap["acl_white_list"] != nil {
-				listener.AclWhiteList = common.StringPtrs(strings.Split(dataMap["acl_white_list"].(string), ","))
-			}
-			if dataMap["client_timeout"] != nil {
-				listener.ClientTimeout = common.StringPtr(dataMap["client_timeout"].(string))
-			}
-			if dataMap["client_timeout_unit"] != nil {
-				listener.ClientTimeoutUnit = common.StringPtr(dataMap["client_timeout_unit"].(string))
-			}
-			if dataMap["connect_timeout"] != nil {
-				listener.ConnectTimeout = common.StringPtr(dataMap["connect_timeout"].(string))
-			}
-			if dataMap["connect_timeout_unit"] != nil {
-				listener.ConnectTimeoutUnit = common.StringPtr(dataMap["connect_timeout_unit"].(string))
-			}
-			if dataMap["listener_mode"] != nil {
-				listener.ListenerMode = common.StringPtr(dataMap["listener_mode"].(string))
-			}
-			if dataMap["listener_name"] != nil {
-				listener.ListenerName = common.StringPtr(dataMap["listener_name"].(string))
-			}
-			if dataMap["listener_port"] != nil {
-				listener.ListenerPort = common.IntPtr(dataMap["listener_port"].(int))
-			}
-			if dataMap["max_conn"] != nil {
-				listener.MaxConn = common.IntPtr(dataMap["max_conn"].(int))
-			}
-			if dataMap["scheduler"] != nil {
-				listener.Scheduler = common.StringPtr(dataMap["scheduler"].(string))
-			}
-			if dataMap["server_timeout"] != nil {
-				listener.ServerTimeout = common.StringPtr(dataMap["server_timeout"].(string))
-			}
-			if dataMap["server_timeout_unit"] != nil {
-				listener.ServerTimeoutUnit = common.StringPtr(dataMap["server_timeout_unit"].(string))
-			}
-			listeners = append(listeners, listener)
+	for _, inter := range strategies {
+		strategy, ok := inter.(map[string]interface{})
+		if !ok {
+			break
 		}
-		request.TcpListeners = listeners
+
+		if inter, ok := strategy["http_listeners"]; ok {
+			datas := inter.([]interface{})
+			listeners := make([]*haproxy.DescribeLoadBalancerStrategysHttpListeners, 0)
+			for _, data := range datas {
+				if data == nil {
+					continue
+				}
+				dataMap := data.(map[string]interface{})
+
+				backendServers := make([]*haproxy.DescribeLoadBalancerStrategysBackendServer, 0)
+				if dataMap["backend_server"] != nil {
+					backendServersDatas := dataMap["backend_server"].([]interface{})
+					for _, backendServersData := range backendServersDatas {
+						backendServerMap := backendServersData.(map[string]interface{})
+						backendServers = append(backendServers, &haproxy.DescribeLoadBalancerStrategysBackendServer{
+							IP:      common.StringPtr(backendServerMap["ip"].(string)),
+							MaxConn: common.IntPtr(backendServerMap["max_conn"].(int)),
+							Port:    common.IntPtr(backendServerMap["port"].(int)),
+							Weight:  common.StringPtr(backendServerMap["weight"].(string)),
+						})
+					}
+				}
+
+				certificateIds := make([]*haproxy.DescribeLoadBalancerStrategysCertificateIds, 0)
+				if dataMap["certificate_ids"] != nil {
+					certificateIdsDatas := dataMap["certificate_ids"].([]interface{})
+					for _, certificateIdsData := range certificateIdsDatas {
+						certificateIdsMap := certificateIdsData.(map[string]interface{})
+						certificateIds = append(certificateIds, &haproxy.DescribeLoadBalancerStrategysCertificateIds{
+							CertificateId:   common.StringPtr(certificateIdsMap["certificate_id"].(string)),
+							CertificateName: common.StringPtr(certificateIdsMap["certificate_name"].(string)),
+						})
+					}
+				} else {
+					certificateIds = []*haproxy.DescribeLoadBalancerStrategysCertificateIds{}
+				}
+
+				listener := &haproxy.DescribeLoadBalancerStrategysHttpListeners{
+					BackendServer:  backendServers,
+					CertificateIds: certificateIds,
+				}
+
+				if dataMap["acl_white_list"] != nil {
+					listener.AclWhiteList = common.StringPtrs(strings.Split(dataMap["acl_white_list"].(string), ","))
+				} else {
+					listener.AclWhiteList = []*string{}
+				}
+				if dataMap["client_timeout"] != nil {
+					listener.ClientTimeout = common.StringPtr(dataMap["client_timeout"].(string))
+				}
+				if dataMap["client_timeout_unit"] != nil {
+					listener.ClientTimeoutUnit = common.StringPtr(dataMap["client_timeout_unit"].(string))
+				}
+				if dataMap["connect_timeout"] != nil {
+					listener.ConnectTimeout = common.StringPtr(dataMap["connect_timeout"].(string))
+				}
+				if dataMap["connect_timeout_unit"] != nil {
+					listener.ConnectTimeoutUnit = common.StringPtr(dataMap["connect_timeout_unit"].(string))
+				}
+				if dataMap["listener_mode"] != nil {
+					listener.ListenerMode = common.StringPtr(dataMap["listener_mode"].(string))
+				}
+				if dataMap["listener_name"] != nil {
+					listener.ListenerName = common.StringPtr(dataMap["listener_name"].(string))
+				}
+				if dataMap["listener_port"] != nil {
+					listener.ListenerPort = common.IntPtr(dataMap["listener_port"].(int))
+				}
+				if dataMap["max_conn"] != nil {
+					listener.MaxConn = common.IntPtr(dataMap["max_conn"].(int))
+				}
+				if dataMap["scheduler"] != nil {
+					listener.Scheduler = common.StringPtr(dataMap["scheduler"].(string))
+				}
+				if dataMap["server_timeout"] != nil {
+					listener.ServerTimeout = common.StringPtr(dataMap["server_timeout"].(string))
+				}
+				if dataMap["server_timeout_unit"] != nil {
+					listener.ServerTimeoutUnit = common.StringPtr(dataMap["server_timeout_unit"].(string))
+				}
+				if dataMap["sticky_session"] != nil {
+					listener.StickySession = common.StringPtr(dataMap["sticky_session"].(string))
+				}
+				listeners = append(listeners, listener)
+			}
+			request.HttpListeners = listeners
+		}
+
+		if inter, ok := strategy["tcp_listeners"]; ok {
+			datas := inter.([]interface{})
+			listeners := make([]*haproxy.DescribeLoadBalancerStrategysTcpListeners, 0)
+			for _, data := range datas {
+				if data == nil {
+					continue
+				}
+				dataMap := data.(map[string]interface{})
+
+				backendServers := make([]*haproxy.DescribeLoadBalancerStrategysBackendServer, 0)
+				if dataMap["backend_server"] != nil {
+					backendServersDatas := dataMap["backend_server"].([]interface{})
+					for _, backendServersData := range backendServersDatas {
+						backendServerMap := backendServersData.(map[string]interface{})
+						backendServers = append(backendServers, &haproxy.DescribeLoadBalancerStrategysBackendServer{
+							IP:      common.StringPtr(backendServerMap["ip"].(string)),
+							MaxConn: common.IntPtr(backendServerMap["max_conn"].(int)),
+							Port:    common.IntPtr(backendServerMap["port"].(int)),
+							Weight:  common.StringPtr(backendServerMap["weight"].(string)),
+						})
+					}
+				}
+
+				listener := &haproxy.DescribeLoadBalancerStrategysTcpListeners{
+					BackendServer: backendServers,
+				}
+
+				if dataMap["acl_white_list"] != nil {
+					listener.AclWhiteList = common.StringPtrs(strings.Split(dataMap["acl_white_list"].(string), ","))
+				}
+				if dataMap["client_timeout"] != nil {
+					listener.ClientTimeout = common.StringPtr(dataMap["client_timeout"].(string))
+				}
+				if dataMap["client_timeout_unit"] != nil {
+					listener.ClientTimeoutUnit = common.StringPtr(dataMap["client_timeout_unit"].(string))
+				}
+				if dataMap["connect_timeout"] != nil {
+					listener.ConnectTimeout = common.StringPtr(dataMap["connect_timeout"].(string))
+				}
+				if dataMap["connect_timeout_unit"] != nil {
+					listener.ConnectTimeoutUnit = common.StringPtr(dataMap["connect_timeout_unit"].(string))
+				}
+				if dataMap["listener_mode"] != nil {
+					listener.ListenerMode = common.StringPtr(dataMap["listener_mode"].(string))
+				}
+				if dataMap["listener_name"] != nil {
+					listener.ListenerName = common.StringPtr(dataMap["listener_name"].(string))
+				}
+				if dataMap["listener_port"] != nil {
+					listener.ListenerPort = common.IntPtr(dataMap["listener_port"].(int))
+				}
+				if dataMap["max_conn"] != nil {
+					listener.MaxConn = common.IntPtr(dataMap["max_conn"].(int))
+				}
+				if dataMap["scheduler"] != nil {
+					listener.Scheduler = common.StringPtr(dataMap["scheduler"].(string))
+				}
+				if dataMap["server_timeout"] != nil {
+					listener.ServerTimeout = common.StringPtr(dataMap["server_timeout"].(string))
+				}
+				if dataMap["server_timeout_unit"] != nil {
+					listener.ServerTimeoutUnit = common.StringPtr(dataMap["server_timeout_unit"].(string))
+				}
+				listeners = append(listeners, listener)
+			}
+			request.TcpListeners = listeners
+		}
 	}
 
 	return request
-}
-
-type HaproxyStrategyHttpListenerProviderInput struct {
-	ServerTimeoutUnit  string `json:"server_timeout_unit"`
-	ServerTimeout      string `json:"server_timeout"`
-	StickySession      string `json:"sticky_session"`
-	AclWhiteList       string `json:"acl_white_list"`
-	ListenerMode       string `json:"listener_mode"`
-	MaxConn            int    `json:"max_conn"`
-	Scheduler          string `json:"scheduler"`
-	ConnectTimeout     string `json:"connect_timeout"`
-	ConnectTimeoutUnit string `json:"connect_timeout_unit"`
-	ClientTimeout      string `json:"client_timeout"`
-	ClientTimeoutUnit  string `json:"client_timeout_unit"`
-	ListenerPort       int    `json:"listener_port"`
-	ListenerName       string `json:"listener_name"`
-	BackendServer      []*struct {
-		IP      string `json:"ip"`
-		MaxConn int    `json:"max_conn"`
-		Port    int    `json:"port"`
-		Weight  string `json:"weight"`
-	} `json:"backend_server"`
-	CertificateIds []*struct {
-		CertificateId   string `json:"certificate_id"`
-		CertificateName string `json:"certificate_name"`
-	} `json:"certificate_ids"`
-}
-
-type HaproxyStrategyTcpListenerProviderInput struct {
-	ServerTimeoutUnit  string `json:"server_timeout_unit"`
-	ServerTimeout      string `json:"server_timeout"`
-	AclWhiteList       string `json:"acl_white_list"`
-	ListenerMode       string `json:"listener_mode"`
-	MaxConn            int    `json:"max_conn"`
-	Scheduler          string `json:"scheduler"`
-	ConnectTimeout     string `json:"connect_timeout"`
-	ConnectTimeoutUnit string `json:"connect_timeout_unit"`
-	ClientTimeout      string `json:"client_timeout"`
-	ClientTimeoutUnit  string `json:"client_timeout_unit"`
-	ListenerPort       int    `json:"listener_port"`
-	ListenerName       string `json:"listener_name"`
-	BackendServer      []*struct {
-		IP      string `json:"ip"`
-		MaxConn int    `json:"max_conn"`
-		Port    int    `json:"port"`
-		Weight  string `json:"weight"`
-	} `json:"backend_server"`
 }
