--- examples/cds_haproxy/.terraform.lock.hcl
@@ -1,10 +1,0 @@
-# This file is maintained automatically by "terraform init".
-# Manual edits may be lost in future updates.
-
-provider "terraform.capitalonline.net/capitalonline/cds" {
-  version     = "1.0.0"
-  constraints = "1.0.0"
-  hashes = [
-    "h1:NHm65md+H+uExCxOZGaCwa6diuZhg5Eu2tQW+AIVP6s=",
-  ]
-}
