--- examples/cds_mysql/main.tf
@@ -1,4 +1,0 @@
-data cds_data_source_mysql my_mysql {
-    region_id = "CN_Beijing_E"
-    result_output_file = "data.json"
-}
