--- examples/cds_mysql/versions.tf
@@ -1,8 +1,0 @@
-terraform {
-    required_providers {
-        cds = {
-            source = "terraform.capitalonline.net/capitalonline/cds"
-            version = "1.0.0"
-        }
-    }
-}
