--- examples/cds_vdc/.terraform.lock.hcl
@@ -1,10 +1,0 @@
-# This file is maintained automatically by "terraform init".
-# Manual edits may be lost in future updates.
-
-provider "terraform.capitalonline.net/capitalonline/cds" {
-  version     = "1.0.0"
-  constraints = "1.0.0"
-  hashes = [
-    "h1:WgAfC1d7ymmhrYxyb8+j49xgdNoYvfr9q/ZRi2CXDZ4=",
-  ]
-}
