+++ examples/cds_vdc_private_subnet/variables.tf
@@ -1,0 +1,4 @@
+variable "vdc_name" {
+  //  description = "The vdc name used to launch a new vdc."
+  default = "TF-testAccVdc"
+}
