+++ go.mod
@@ -4,7 +4,7 @@ go 1.13
 
 require (
 	github.com/athom/goset v0.0.0-20151103071336-f07be69aa101
-	github.com/capitalonline/cds-gic-sdk-go v0.2.5
+	github.com/capitalonline/cds-gic-sdk-go v0.2.4
 	github.com/hashicorp/terraform-plugin-sdk v1.14.0
 	github.com/likexian/gokit v0.20.16
 	github.com/terraform-providers/terraform-provider-tencentcloud v1.21.2
